versao modularizada
