+++ ouvICEx/init_flask.py
@@ -3,10 +3,12 @@ from flask_sqlalchemy import SQLAlchemy
 import datetime
 from scripts import analyses as anl
 from scripts.history import app_history
+from scripts.analyses import app_analyses
 from scripts.database import db, posts
 
 app = Flask(__name__)
 app.register_blueprint(app_history, url_prefix="")
+app.register_blueprint(app_analyses, url_prefix="")
 app.config['SQLALCHEMY_TRACK_NOTIFICATIONS'] = False
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///posts.sqlite3'
 
@@ -34,19 +36,6 @@ def form():
         return render_template("form.html")
     else:
         return render_template("form.html")
-
-@app.route("/analyses", methods=["POST", "GET"])
-def analyses():
-    #print("entrou")
-    #anl.test()
-    #return "analyses content"
-    if request.method == "GET":
-        return render_template(
-                    "analyses.html",
-                    values = posts.query.all()
-                )
-    else:
-        return "ERROR"
 
 if __name__ == "__main__":
